+++ src/Core/Plugin.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
  */
 namespace MeTools\Core;
 
-use Cake\Core\Exception\MissingPluginException;
 use Cake\Core\Plugin as CakePlugin;
 use Tools\Exceptionist;
 use Tools\Filesystem;
@@ -59,7 +58,7 @@ class Plugin extends CakePlugin
      * @param string|null $file File
      * @param bool $check Checks if the file exists
      * @return string Path of the plugin or path of the path of a plugin file
-     * @throws \Throwable
+     * @throws \Tools\Exception\FileNotExistsException|\ErrorException
      */
     public static function path(string $name, ?string $file = null, bool $check = false): string
     {
@@ -69,7 +68,9 @@ class Plugin extends CakePlugin
         }
 
         $path = $plugin . $file;
-        Exceptionist::isTrue(is_readable($path) || !$check, __d('me_tools', 'File or directory `{0}` does not exist', Filesystem::instance()->rtr($path)), MissingPluginException::class);
+        if ($check) {
+            Exceptionist::fileExists($path, __d('me_tools', 'File or directory `{0}.{1}` does not exist', $name, Filesystem::instance()->rtr($path)));
+        }
 
         return $path;
     }