+++ CHANGELOG.md
@@ -10,7 +10,15 @@
 
 ## 2.21 branch
 ### 2.21.5
-* updated for php-tools 1.6.5.
+* added `iframe()` and `youtube()` methods to the `BootstrapHtmlHelper`;
+* the `HtmlHelper` is now totally deprecated, use `BootstrapHtmlHelper` instead. In a later version, the latter
+  will take the place of the former, assuming its name;
+* the `DropdownHelper` is now totally deprecated, use `BootstrapDropdownHelper` instead. In a later version, the latter
+  will take the place of the former, assuming its name;
+* the `FormHelper` is now totally deprecated, use `BootstrapFormHelper` instead. In a later version, the latter will
+  take the place of the former, assuming its name;
+* improved \MeTools\Core\Plugin::path()` method;
+* updated for php-tools 1.6.5 and 1.7.0.
 
 ### 2.21.4
 * `append-text` and `prepend-text` options for `BootstrapFormHelper::control()`
@@ -293,7 +301,7 @@
 * added `TestCaseTrait::assertIsMock()` assert method.
 
 ### 2.17.4
-* fixed bug in the integration with CKeditor when uploading images (see
+* fixed bug in the integration with CKEditor when uploading images (see
     [this issue](https://github.com/sunhater/kcfinder/issues/171)).
 
 ### 2.17.3
@@ -675,7 +683,7 @@
 * added support for Shareaholic;
 * jQuery, Bootstrap, Bootstrap Date/Time Picker, Moment.js and Font Awesome
 	are installed via Composer;
-* improved the logs management;
+* improved the log management;
 * fixed little bugs.
 
 ## 2.0 branch